inventory saves slot as byte, added beginning of spell functionality

+++ src/main/java/com/xylo_datapacks/energy_manipulation/util/InventoryUtils.java
@@ -14,7 +14,7 @@ public class InventoryUtils {
             ItemStack stack = inventory.getStack(i);
             if (!stack.isEmpty()) {
                 NbtCompound stackTag = stack.writeNbt(new NbtCompound());
-                stackTag.putInt("Slot", i);
+                stackTag.putByte("Slot", (byte) i); 
                 tag.add(stackTag);
             }
         }
